surge(invest)
--> this is surge module and it uses for investment. in it, every partner can get own scheme. and showing own scheme with help of api.

-->1. controller (surgeadmin.php - copied)
-->2.model(Investmodel.php - copied)
                  (Surge_model - copied)
-->3.view(all view files copied)

+++ application/modules/surgeadmin/views/buisnessteam/allrepersentative.php
@@ -1,3 +1,4 @@
+
 <?php 
 if($this->session->userdata('session_data'))
 {
@@ -6,54 +7,72 @@ if($this->session->userdata('session_data'))
 }
 else
 {
-    redirect(base_url('Surgeadmin/businessTeamlogin'));
+    redirect(base_url('Surgeadmin/login'));
 }
  ?>
 
+<link rel="stylesheet" href="https://cdn.datatables.net/1.13.4/css/dataTables.bootstrap.min.css">
+<script src="https://cdn.datatables.net/1.13.4/js/jquery.dataTables.min.js"></script>
+<script src="https://cdn.datatables.net/1.13.4/js/dataTables.bootstrap.min.js"></script>
+<script>
+    $(document).ready(function() {
+
+        $('#redemption').DataTable();
+    });
+</script>
+
+<div class="row">
+<div class="col-md-12">
 <div class="box-body table-responsive no-padding">
-    <center><h3 style="border-bottom:5px solid blue;">All Representative</h3></center>
+<div style="float:right; padding: 5px;"><button type="button"><a href="Surgeadmin/reg_representative">+Add Repersentative</a></button></div>
+    <h3 style="text-align:center;">All Representative</h3>
 
     <?php if($this->session->flashdata('flashmsg')): ?>
-        <p style="background-color:yellow; font-size:medium;"><?php echo $this->session->flashdata('flashmsg'); unset($_SESSION['flashmsg']); ?></p>
+        <p style="background-color:yellow; font-size:medium;"><?php echo $this->session->flashdata('flashmsg'); unset($_SESSION['flashmsg']) ?></p>
            <?php endif ?>
-
-              <table id="test2" class="table table-bordered table-hover">
+        
+              <table id="redemption" class="table table-bordered table-hover">
+                <thead>
               <tr>
                 <th>Rid</th>
               <th>RepName</th>
-            <th>vender</th>
+            <th>Partner Name</th>
             <th>RepDesignation</th>
             <th>Repphone</th>
             <th>Repemail</th>
-            <th>password</th>
             <th>created_date</th>
             <th>Action</th>
+            <th></th>
         </tr>
+                </thead>
+                <tbody>
         <?php      
    if(@$row)
    {
- 
+    $i=1;
 foreach($row as $row1)   { ?>
 
         <tr>
-            <td><?php echo $row1['rid']; ?></td>
+            <td><?php echo $i; ?></td>
             <td><?php echo $row1['RepName']; ?></td>
             <td><?php echo $row1['Company_Name']; ?></td>
             <td><?php echo $row1['RepDesignation']; ?></td>
             <td><?php echo $row1['Repphone']; ?></td>
             <td><?php echo $row1['Repemail']; ?></td>
-            <td><?php echo $row1['password']; ?></td>
             <td><?php echo $row1['created_date']; ?></td>
             <td><a class="btn btn-success" href="<?= base_url(); ?>Surgeadmin/editrepersent/<?= $row1['rid']; ?>"> Edit</a></td>
-<td><a class="btn btn-danger" href="<?= base_url(); ?>Surgeadmin/deleterepersent/<?=$row1['rid']; ?>">Delete</a></td>
+            <td><a class="btn btn-danger" href="<?= base_url(); ?>Surgeadmin/deleterepersent/<?=$row1['rid']; ?>" class="btn btn-danger" onclick="return confirm('Are You Sure Want To Delete This')">Delete</a></td>
 
              </tr>
-        <?php } }
+        <?php $i++;
+         } }
         
         else { echo "Sorry There are no redemtion requests yet!"; } ?>
+                </tbody>
 </table>
-        </div>
-        <?php
-						if($this->session->flashdata('redeemfinal')) {	?>
-						 <h4 class="text-success text-center" style="margin-top: 10px;"> <?=$this->session->flashdata('redeemfinal')?></h4>
-						<?php } ?>
+        </div><div class="clearfix"></div>
+       </div>
+</div>
+
+
+
